Created Unit base class and a couple of units inheriting from the base class
Look this over and let me know if this is the implementation you want to go with. They different units inherit from the base units.

To create a new unit you must create a new inheriting scene to inherit from Unit.tscn. Then rename the "Unit" node to the name of the chess piece. Then you have to go into its UnitIcon and set its sprite frames. In the sprite sheet, one animation must be called white, and one must be called black.