Changes to atom highlight method, fixed z index, made UI changes for units
Added subtle movement highlights and methods for it in Tile.

Fixed the z index problem for sure this time. I think.

Changed the way atoms are processed. There is now a variable called mode. For each direction this mode starts with value "active". While active any tiles processed are added to possible positions. When impassible terrain is found it switches to passive mode. Any nodes found from here on out are added to subtle_possible_positions.

This makes it so that the movement and attack pattern is more clear while still keeping only certain tiles as possible moves/attacks.